pyproject.toml
[tool.poetry]
name = "websockets-demo"
version = "0.1.0"
description = ""
authors = ["balo <98645847+balojey@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
